Rename api packages for Modulith compatibility (markapi->mark.api, fileapi->file.api)
