+++ file-api/src/main/java/pt/estga/file/api/FileStorageOperations.java
@@ -1,4 +1,4 @@
-package pt.estga.fileapi;
+package pt.estga.file.api;
 
 import pt.estga.file.dtos.MediaFileDto;
