+++ file-api/src/main/java/pt/estga/file/api/StagedFileRecord.java
@@ -1,4 +1,4 @@
-package pt.estga.fileapi;
+package pt.estga.file.api;
 
 import java.util.UUID;
