+++ mark-api/src/main/java/pt/estga/mark/api/MarkEvidenceQueryService.java
@@ -1,11 +1,10 @@
-package pt.estga.markapi;
+package pt.estga.mark.api;
 
 import pt.estga.mark.dtos.EvidenceMarkDto;
 import pt.estga.mark.dtos.MarkEvidenceDistanceDto;
 import pt.estga.mark.dtos.MarkEvidenceDto;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 public interface MarkEvidenceQueryService {
