+++ mark/src/main/java/pt/estga/mark/services/MarkEvidenceQueryServiceImpl.java
@@ -9,7 +9,7 @@ import pt.estga.mark.dtos.MarkEvidenceDistanceDto;
 import pt.estga.mark.dtos.MarkEvidenceDto;
 import pt.estga.mark.mappers.MarkEvidenceMapper;
 import pt.estga.mark.repositories.MarkEvidenceRepository;
-import pt.estga.markapi.MarkEvidenceQueryService;
+import pt.estga.mark.api.MarkEvidenceQueryService;
 
 import java.util.List;
 import java.util.UUID;
